timing is cleaned up for all of that stuff, prelude text now has the proper line breaks too

+++ src/main/Middlemarch/NovelTextComponent.h
@@ -1,8 +1,10 @@
 #ifndef TEXTBASED_NOVELTEXTCOMPONENT_H
 #define TEXTBASED_NOVELTEXTCOMPONENT_H
+#include <string>
 
 struct NovelTextComponent {
-    int readIndex = 0;
+    size_t readIndex = 0;
+
 };
 
 #endif
